Remove validation from the controller
Don't do validation in the controller as that just throws an exception
without pointing out where the issue is.

Now the validation of the dates is done in the model and it enforces the
date format to be in dd/mm/yyyy format because if somebody specifies
dd/mm/yy for example 10/10/80 the yy part is parsed as '0080'.