Created HealthStatus controller & check
Re-named Ping controller to HealthStatus, because '/ping.json' &
'/healthcheck.json' will preform pretty much a similar functionality, so
it makes sense to have just one controller for both.

Added the '/healthcheck.json' endpoint, which for now, will just return
the database info. The plan is to add the checks for more services.