Faster TravisCI builds
When building jobs on Travis CI, this message shows up:

`This job ran on our legacy infrastructure. Please read our docs on how
to upgrade'

The link to the document is: http://j.mp/1CtCCRO

Trying it out with this commit.

But still not sure if I have to remove all the mentions of `sudo' from
.travis.yml.