Handle personal information for form object
This is a dirty workaround, as we're still mid flight. The idea is to
render the form object for the initial page, which is for now only
personal information. But the plan is to move all the pages to this. So
the controller will get the proper attention it deserves. For now giving
it extra baggage which WILL get refactored.