Handle formatting of NI number in personal details
These checks exist in the previous version of the monolithic application
model. They were only caught by our managers who were inputing the test
data in manually and typed in their NI numbers in lower case.

Implemented the formatting as the pre validation callback. The
functionality is not significant enough to do this in a separate class
some cleaner design pattern.