Introduced a new namespace under models
Since there will be two major flows:

1. for Application processing
2. for Evidence processing

there will be different form objects that handle their forms. So
introducing namespacing in order to organise the code, so that anybody
looking at the codebase can know where the models for Application sit
and where models for Evidence sit.

Used the name 'Applikation' as it was conflicting with 'Application' and
Rails server would not even start.