Removed the formatting callback for Income
In Income form object, there was a pre validation callback, which
prevented the validation errors surfacing to the view when the user
entered nothing or non integer values. By removing the callback, which
was planned regardless, the errors now surface to the model and the user
is prompted to enter the correct income amount.