Formatting of the Income amount revisited
Previously the `format_amount' method only rounded the amount supplied,
and then the amount needed to be turned into integer before being
supplied to income calculation. But since `format_amount' is already
doing the formatting it might aswell turn the `amount' attribute into
the valid type that can be used by the income calculator. The only
reason it was set as a string is so that the formatting can be done on
decimal numbers without losing the value provided after the decimal
point.