Tweaked the confirmation views more
Added the appropriate copy to the partials that are meant to render when
the remission is partial or there is no remission.

This still needs a bit of work, for example passing in the values for
the locales so that the templates are populated correctly. I plan on
doing this via the view object as it was done previously for other
views.