Don't remove the reason explanation
If the user chooses the option that the evidence check does not pass and
they fill out the reason, but they then toggle the option and come back
to it, the reason explanation should not be removed.

This came out of the discussion with the team, as this way of
functioning was perceived as more helpful to the end user.