Implemented application emergency
Some application's will need to be expedited through the system, so
they'll need to be declared as emergency applications.

This change allows for that.

WRONG THINGS: currently in the ApplicationDetail model I'm assigning an
attribute called `emergency' which I manually remove in the Application
controller and in the ApplicationDetail model specs. The reason for that
is that currently Application model is one huge monolith and once we
split it up, these ugly hacks will be removed.