Added a callback for evidence_reason
The plan is to migrate away from this monolithic Application model and
process individual pages/steps of the application processing with the
form objects. Which is already being done. But the persistance is still
handled by the monolithic Application model.

With the refactoring that is due soon, this ugly hack will be addressed
and thus the callback will be removed. Adding it now just to get the
feature in.