When no NI number provided, prevent loop
There was an issue when a user didn't provide their NI number and the
application was started, the benefit check was attempted but it could
not be done because benefit check requires the NI number for the lookup.
This would throw the user into a loop. This change tries to prevent
that.

Most of the changes in this commit are related to the renaming of the
buttons that need to be clicked to progress through the application
flow.