Add factory for Applicant and make it work
It's quite tricky, because Application factory needs Applicant factory
and vice versa. So there's a trick using after(:build) callbacks in both
factories to make that possible.
The reason for doing all this is to not having to change all existing
tests. But eventually when Application is a very thin model, hopefully
these hacks can go away.