Added 'outcome' to PartPayment
This was added so that the ResolverService can be used to set the state
of the PartPayment & Application's decision in one swoop. Also, this is
the way it was implemented for EvidenceCheck and the intention is to
have all the objects which deal with the Application's decision, to be
processed this way.