Change status migration to migrate wrong data too
Only applications which have completed_at column set should be
considered as processed. However the current user journey allows for
users not to click the last button to finish their job. That problem has
been fixed in a different piece of work, but this one has to take into
account the data which had been created in the meantime.