Allow ResolverService to complete records
This now allows Application, EvidenceCheck and PartPayment to be
completed. It also always updates the relevant Application and if
needed, stores the final decision.

Unfortunatelly I just discovered that PartPayment won't have an outcome
set from the normal journey, which has to be sorted out. It'll only have
outcome for return. But that means that trying to complete a PartPayment
using the ResolverService will fail.