Implement frontend for removing application form
It uses the html5 details tag. When there's an error and the page is
re-rendered, the details panel with the form is automatically expanded.
Tested in Chrome and Firefox. When this situation occurs in Firefox, the
triangle icon is not correctly rotated, but I think it's too minor to
spend long time debugging.