Authorise ProcessedApplications#update
There's a little issue with our current setup of form objects. They also
have responsibility to save/update the objects, but that makes it
impossible to authorise the changed object. We'll have to somehow deal
with it, but for now we authorise the object itself before any changes
to it are made.