Start now only appears when jurisdictions assigned
It can happen that an office has no jurisdictions assigned and then when
users try to process an application, it blows in later stage. To prevent
that, we won't show the Start now button until we know they can safely
process applications.

It's not forbidden by the policy though, so theoretically somebody could
bypass it by POSTing to the /applications/create url, but then it's
their problem.