Move BEC edit link from office edit to show
It's mainly because of the permissions need to be separated for
different types of users.

On the homepage, the reports and offices boxes were moved down the page.
But there is an issue on the homepage, which is that we use header html
tag multiple times, which should be sorted out.