Fix emergency fields load/save
Because we only store emergency_reason in the database, if the page is
visited in the edit mode, the 'emergency' flag was not set, but the
reason was present. So the user doesn't see that the application is
setup as emergency. And most importantly they weren't able to change it
back to non-emergency.