Don't create BC just to check status
If the BenefitCheck was not created in in #run method then it was created in can_override? check. Which is something that we don't want. If the check exist I jus load the previous one if not then there is a reason for it, so I'm prevening it to create it.