RST-1845 - Search query update (#897)
* Refactoring search

The search result should not be limited by the office. To improve the speed after extending search scope I use direct search for reference numbers and new search query just for names.

* Adding the office filter back for users only

User doesn't need to see other officess applications. This is just for admins only. It will help with the search speed for normal users.

* Updating search query logic

When the search for reference number returns an application from different office it displays error.

* RST-1845 cucumber test for searching for an application for another office